Updated script_17_FA.py
Ensured that the script was reading the correct file. Added a line to print the head of the table to preview. Commented the 17 variables that are needed to call createStudent from init_database.py and what is included in the csv file. Will need to restructure either the createStudent method or the csv files.